# apps/quant-console/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CompareRequest": {
        "additionalProperties": false,
        "description": "Experiments selected for a compatibility check.",
        "properties": {
          "artifact_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 4,
            "minItems": 2,
            "title": "Artifact Ids",
            "type": "array"
          }
        },
        "required": [
          "artifact_ids"
        ],
        "title": "CompareRequest",
        "type": "object"
      },
...
      "CompatibilityResult": {
        "additionalProperties": false,
        "description": "Comparison permission and its explicit limitations.",
        "properties": {
          "changed_fields": {
            "items": {
              "type": "string"
            },
            "title": "Changed Fields",
            "type": "array"
          },
          "economic_inference_allowed": {
            "title": "Economic Inference Allowed",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExperimentDetail"
            },
            "title": "Items",
            "type": "array"
          },
          "mode": {
            "$ref": "#/components/schemas/ComparisonMode"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          }
        },
        "required": [
          "snapshot_id",
          "mode",
          "reasons",
          "changed_fields",
          "economic_inference_allowed",
          "items"
        ],
        "title": "CompatibilityResult",
        "type": "object"
      },
...
      "DataEvaluability": {
        "description": "Whether absolute experiment metrics can be evaluated.",
        "enum": [
          "VALID",
          "NOT_EVALUABLE",
          "UNKNOWN"
        ],
        "title": "DataEvaluability",
        "type": "string"
      },
      "EvidenceSummary": {
        "additionalProperties": false,
        "description": "Safe source identity without local paths or restricted payloads.",
        "properties": {
          "absolute_paths_exposed": {
            "default": false,
            "title": "Absolute Paths Exposed",
            "type": "boolean"
          },
          "calendar_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendar Sha256"
          },
          "data_use_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Use Level"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "receipt_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Receipt Sha256"
          },
          "rules_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rules Sha256"
          },
          "run_identity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Identity"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "source_kind": {
            "title": "Source Kind",
            "type": "string"
          },
          "study_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Id"
          },
          "study_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Revision"
          },
          "trading_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trading Date"
          }
        },
        "required": [
          "evidence_id",
          "source_kind"
        ],
        "title": "EvidenceSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SeriesResponse": {
        "additionalProperties": false,
        "description": "Full precision series bound to one snapshot.",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/SeriesMeta"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/SeriesPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          }
        },
        "required": [
          "snapshot_id",
          "meta",
          "points"
        ],
        "title": "SeriesResponse",
        "type": "object"
      },
...
      "SortDirection": {
        "description": "Stable sort direction.",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortDirection",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Quant Console V1.5",
    "version": "1.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/snapshots/{snapshot_id}/comparisons/validate": {
      "post": {
        "operationId": "validate_comparison_api_v1_snapshots__snapshot_id__comparisons_validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompatibilityResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Comparison"
      }
    },
    "/api/v1/snapshots/{snapshot_id}/evidence/{evidence_id}": {
      "get": {
        "operationId": "evidence_api_v1_snapshots__snapshot_id__evidence__evidence_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Snapshot Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "pattern": "^[A-Za-z0-9_.:-]+$",
              "title": "Evidence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evidence"
      }
    },
    "/api/v1/snapshots/{snapshot_id}/experiments": {
      "get": {
        "operationId": "experiments_api_v1_snapshots__snapshot_id__experiments_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Snapshot Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "study",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Study"
            }
          },
          {
            "in": "query",
            "name": "family",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Family"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DataEvaluability"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "economic_outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Economic Outcome"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 10000,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "experiment_id",
              "maxLength": 64,
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortDirection",
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Experiments"
      }
    },
    "/api/v1/snapshots/{snapshot_id}/experiments/{artifact_id}": {
      "get": {
        "operationId": "experiment_detail_api_v1_snapshots__snapshot_id__experiments__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Snapshot Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "pattern": "^[A-Za-z0-9_.:-]+$",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Experiment Detail"
      }
    },
    "/api/v1/snapshots/{snapshot_id}/exports/experiments": {
      "post": {
        "operationId": "export_experiments_api_v1_snapshots__snapshot_id__exports_experiments_post",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Experiments"
      }
    },
...
    "/api/v1/snapshots/{snapshot_id}/overview": {
      "get": {
        "operationId": "overview_api_v1_snapshots__snapshot_id__overview_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Overview"
      }
    },
    "/api/v1/snapshots/{snapshot_id}/prospective": {
      "get": {
        "operationId": "prospective_api_v1_snapshots__snapshot_id__prospective_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProspectiveSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prospective"
      }
    },
    "/api/v1/snapshots/{snapshot_id}/series/{series_id}": {
      "get": {
        "operationId": "series_api_v1_snapshots__snapshot_id__series__series_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Snapshot Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Series Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Series"
      }
    },
    "/api/v1/snapshots/{snapshot_id}/signals": {
      "get": {
        "operationId": "signals_api_v1_snapshots__snapshot_id__signals_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalsSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signals"
      }
    },
    "/api/v1/snapshots/{snapshot_id}/studies": {
      "get": {
        "operationId": "studies_api_v1_snapshots__snapshot_id__studies_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Studies"
      }
    }
  }
}